+++ ledscape.h
@@ -68,8 +68,8 @@ ledscape_draw(
 extern void
 ledscape_set_color(
 	ledscape_frame_t * const frame,
-	uint8_t strip,
-	uint8_t pixel,
+	uint32_t strip,
+	uint32_t pixel,
 	uint8_t r,
 	uint8_t g,
 	uint8_t b
